+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
-### Changed
+### Removed
 
 - Removed availability zones for `GermanyWestCentral` in `Azure`.
 - Removed the `--domain` flag on `Azure`.