+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [0.10.0] - 2020-10-23
+
 ### Removed
 
 - Removed availability zones for `GermanyWestCentral` in `Azure`.
@@ -127,7 +129,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ## [0.2.0] 2020-04-23
 
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/giantswarm/kubectl-gs/compare/v0.7.2...v0.8.0
 [0.7.2]: https://github.com/giantswarm/kubectl-gs/compare/v0.7.1...v0.7.2